#7 Write translations to resx file as an aditional option to Excel
