+++ ListMissingResourceItems/ListMissingResourceItems.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="System.Linq.Async" Version="6.0.1" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\WriteMissingResourceItems\WriteMissingResourceItems.csproj" />
+  </ItemGroup>
+
 </Project>
